Move key store code to its own file
